==> Lessons/06_Lab_GLMs_Kelly.Rmd <==
---
title: "6: Part 1 - Generalized Linear Models"
author: "Environmental Data Analytics | John Fay and Luana Lima | Developed by Kateri Salk"
date: "Fall 2022"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## Objectives
1. Answer questions on M5/A5
2. Answer questions on M6 - GLMs
3. Practice more application GLM to real datasets


## Set up
```{r, message = FALSE}
library(tidyverse)
install.packages("agricolae")
library(agricolae)   

EPAair <- read.csv("./Data/Processed/EPAair_O3_PM25_NC1819_Processed.csv", stringsAsFactors = TRUE)
# Set date to date format
EPAair$Date <- as.Date(EPAair$Date, format = "%Y-%m-%d")

Litter <- read.csv("./Data/Processed/NEON_NIWO_Litter_mass_trap_Processed.csv", stringsAsFactors = TRUE)
# Set date to date format
Litter$collectDate <- as.Date(Litter$collectDate , format = "%Y-%m-%d")

# Set theme
mytheme <- theme_classic(base_size = 14) +
  theme(axis.text = element_text(color = "black"), 
        legend.position = "top")
theme_set(mytheme)
```

Workflow Methods:
View the data - scatterplot
Create the linear model (lm function)
Examine model summary & create summary plots

High leverage = that data point has a large effect or weight on the coefficients
dont want to have outliers in the top right of the residuals vs. leverage plot
Tells you if outliers are having a large effect on the model

AIC can be negative, but don't need to consider the negative sign - just look at absolute value
The smaller the AIC, the better the model

Multiple Regression Workflow
Generate linear model
Apply step() function to resulting model
Run suggested model and report findings - does r^2 increase? If it does, that means there has been an improvement

ANOVA - predicitng Y from categorical variables - compares means
Factor: a variable used to group data, suspected to explain variability in another [response] variable
Levels: the different values found in the factor
Balanced Design: all levels have equal numbers of observations

T-tests
1-sample & 2-sample
1-sided & 2-sided
Need to create null & alternative hypothesis


## Visualization and interpretation challenge

Create three plots, each with appropriately formatted axes and legends. Choose a non-default color palette.

1. geom_density of ozone divided by year [one line for each year] (distinguish between years by adding transparency to the geom_density layer).
2. geom_boxplot of ozone divided by year. Add letters representing a significant difference between 2018 and 2019 (hint: stat_summary). 
3. geom_violin of ozone divided by year, with the 0.5 quantile marked as a horizontal line. Add letters representing a significant difference between 2018 and 2019. 

```{r}
#Exercise 1:

density <-
  ggplot(EPAair, aes(x = Ozone, fill = as.factor(Year))) +
  geom_density(alpha = 0.5) +
  scale_fill_manual(value = c("dark green" , "dark blue")) +
  label(x = "Ozone AQI Value", y = "Density")
print(density)

boxplot <- 
  ggplot(EPAair, aes(x = Year, y = Ozone)) +
  geom_boxplot() +
  stat_summary(geom = "text", fun = max, vjust = -1, size =4)
print(boxplot)


```

## Linear Regression

Important components of the linear regression are the correlation and the R-squared value. The **correlation** is a number between -1 and 1, describing the relationship between the variables. Correlations close to -1 represent strong negative correlations, correlations close to zero represent weak correlations, and correlations close to 1 represent strong positive correlations. The **R-squared value** is the correlation squared, becoming a number between 0 and 1. The R-squared value describes the percent of variance accounted for by the explanatory variables. 

For the NTL-LTER dataset, can we predict PM2.5 (y-axis) from Ozone (x-axis)?

```{r}

#Exercise 2: Run a linear regression PM2.5 by Ozone. Find the p-value and R-squared value. 

#EPAair <- filter(EPAair, PM2.5 != 0)

PM2.5_Ozone <-
  lm(data = EPAair, PM2.5 ~ Ozone)
summary(PM2.5_Ozone)

#p-value less than 0.5 (very small so reject null hypothesis)
#Null hypothesis: there is no relationship or coefficient is equal to 0
#r^2 = 0.1316 - 13% of the variability in PM2.5 is explained by ozone

#Exercise 3: Build a scatterplot. Add a line and standard error for the linear regression. Add the regression equation to the plot

ex2plot <-
  ggplot(EPAair, aes(x = Ozone, y = PM2.5)) +
  geom_point() +
  geom_smooth(method = 'lm')
print(ex2plot)

#how do we find other significant variables? using AIC
```


## AIC to select variables

What other variables can we add to improve model?

```{r}

#Exercise 4: Build correlation plots and identify more possible explanatory variables to add to the regression.

ex4subset <- 
  EPAair %>% 
  select(meanLatitude:Ozone) %>% 
  na.omit() #remember to ALWAYS drop na's before doing cor function
  
ex4corplot <- cor(ex4subset)
corrplot(ex4corplot, method = "ellipse") #symmetrical matrix so only need to look at either above or below the diagonal
corrplot.mixed(ex4corplot, upper = "ellipse")

#looking at what else correlates with PM2.5 aside from ozone
#PM2.5 & meanLat = positive corr (0.25)


#Exercise 5: Choose a model by AIC in a Stepwise Algorithm. Do the results from AIC match the variables you selected on Exercise 4?

ex5AIC <- lm(data = ex4subset, PM2.5 ~ Ozone + Year + Month + meanLongitude + meanLatitude)
step(ex5AIC)
#AIC is like a scale - is it worth adding another variable to explain the varaiabiliyt while increasing the complicatedness of the model
#AIC suggests we remove Year, that will give us a lower AIC which is ideal!
#AIC might not always make a suggestion of what to remove


#Exercise 6: Run another regression using the variables selected on Exercise 6. Compare r-squared value with the one from Exercise 2. 

ex6 <- lm(data = ex4subset, PM2.5 ~ Ozone + Month + meanLongitude + meanLatitude)
summary(ex6)

#r^2 is higher so that explains more of the variability

```

## Litter Exercise

```{r}
# Wrangle the data
Litter.Totals <- Litter %>%
  group_by(plotID, collectDate, nlcdClass) %>%
  summarise(dryMass = sum(dryMass))

# Format ANOVA as aov
Litter.Totals.anova <- aov(data = Litter.Totals, dryMass ~ plotID)
summary(Litter.Totals.anova)

# Extract groupings for pairwise relationships
Litter.Totals.groups <- HSD.test(Litter.Totals.anova, "plotID", group = TRUE)
Litter.Totals.groups$groups

Litter.Totals <- Litter.Totals %>%
  mutate(treatgroups = Litter.Totals.groups$groups[plotID,2])

# Graph the results
Litter.Totals.plot <- ggplot(Litter.Totals, aes(x = plotID, y = dryMass)) +
  geom_boxplot() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  stat_summary(geom = "text", fun = max, vjust = -1, size = 3.5,
               label = c("abc", "ab", "abcd", "bcd", "bcd", "a", 
                         "d", "abcd", "cd", "d", "abcd", "abcd")) +
  labs(x = "Plot ID", y = "Dry Mass (g)") +
  ylim(0, 85)
print(Litter.Totals.plot)

```

```{r}
#Exercise 7: Improve the plot

#assign.group <- function(site)
  return(Litter.Totals.groups$groups[as.character(site, 2)])

Litter.Totals$plotID <-
  reorder(Litter.Totals$plotID, Litter.Totals$dryMass, FUN = mean)

```
